# restapi/app/src/swagger_output.json
{
  "openapi": "3.1.0",
  "info": {
    "version": "v1.0.0",
    "title": "SakauBelajar",
    "description": "Dokumentasi API SakauBelajar"
  },
  "servers": [
    {
      "url": "http://194.233.93.124:3030",
      "description": "kalau akses pake IP"
    },
    {
      "url": "http://localhost:3030",
      "description": "kalau akses pake forward port (localhost)"
    }
  ],
  "paths": {
...
    "/test/testdata": {
      "get": {
        "tags": [
          "Test Route"
        ],
        "description": "Ini test route buat dapetin data",
        "responses": {
          "200": {
            "description": "OK"
          }
        }
      },
      "post": {
        "tags": [
          "Test Route"
        ],
        "description": "Ini test route buat nambahin data",
        "responses": {
          "200": {
            "description": "OK"
          }
        },
        "requestBody": {
          "content": {
            "application/json": {
              "schema": {
                "type": "object",
                "properties": {
                  "test_text": {
                    "example": "any"
                  },
                  "test_number": {
                    "example": "any"
                  }
                }
              }
            }
          }
        }
      }
    },
    "/test/testdata/{id}": {
      "delete": {
        "tags": [
          "Test Route"
        ],
        "description": "Ini test route buat hapus data",
        "parameters": [
          {
            "name": "id",
            "in": "path",
            "required": true,
            "schema": {
              "type": "string"
            }
          }
        ],
        "responses": {
          "200": {
            "description": "OK"
          }
        }
      }
    },
...
    "/test/testmysql2": {
      "get": {
        "tags": [
          "Test Route"
        ],
        "description": "Test dapetin data auth dimana username yang ada kata yang dicari nya. Contoh: username = sendi",
        "parameters": [
          {
            "name": "username",
            "in": "query",
            "description": "Username yang ingin dicari",
            "schema": {
              "type": "string"
            }
          }
        ],
        "responses": {
          "200": {
            "description": "OK"
          },
          "400": {
            "description": "Bad Request"
          }
        }
      }
    },
...
  }
}